+++ pac4j-core/src/main/java/org/pac4j/core/util/Pac4jConstants.java
@@ -44,6 +44,8 @@ public interface Pac4jConstants {
     /* The element (client or authorizer) separator */
     String ELEMENT_SEPARATOR = ",";
 
+    String ADD_ELEMENT = "+";
+
     String TYPED_ID_SEPARATOR = "#";
 
     /* The logout pattern for url */