+++ documentation/docs/release-notes.md
@@ -17,6 +17,7 @@ title: Release notes&#58;
 - Removed the 307 HTTP code for a new POST request after a POST request (use 200 instead)
 - Turned the `UserProfile` component into a pure interface and use it as much as possible (especially in the `JwtGenerator` and `JwtAuthenticator`)
 - The `ProfileHelper.restoreOrBuildProfile` method has been removed and the behavior is controlled by the `ProfileDefinition` and its `setRestoreProfileFromTypedId` method (enabled for JWT, disabled for others)
+- Authorizers and matchers can be defined additionaly with "+"
 
 **v4.2.0**: